% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_availability.R
\name{check_availability}
\alias{check_availability}
\title{Check which, if any, years a variable is available for from a given ROMS simulation}
\usage{
check_availability(
  var,
  type = c("hindcast", "projection", "historical"),
  scenario = c("SSP126", "SSP585", "RCP45", "RCP85"),
  version = c("K20P19", "K20", "H16"),
  earth_model = c("GFDL", "CESM", "MIROC"),
  print_console = TRUE
)
}
\arguments{
\item{var}{variable name}

\item{type}{"hindcast", "projection", or "historical" run}

\item{scenario}{"SSP126" or "SSP585" for CMIP6 (`version = "K20P19"` or `version = "K20"`) outputs, "RCP45" or "RCP85" for CMIP5 (`version = "H16"`) outputs}

\item{version}{Either (1) "K20P19" for 30-layer ROMs model from Kearney et al. 2020, with carbonate dynamics from
Pilcher et al. 2019 (default), (2) "K20" for 30layer model without carbonate dynamics, or (3) "H16" for 10-layer
ROMS model from Hermann et al. 2016}

\item{earth_model}{"GFDL", "CESM", or "MIROC"}

\item{print_console}{Whether to print available time blocks to the console instead of returning output}
}
\value{
If `print_console = FALSE`, a list of available time blocks and links to datasets for each time block
}
\description{
Check which, if any, years a variable is available for from a given ROMS simulation
}
