% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_correct.R
\name{delta_correct}
\alias{delta_correct}
\title{Delta-correct ROMS level 2 outputs}
\usage{
delta_correct(x, hindcast, historical, lower = NA, upper = NA)
}
\arguments{
\item{x}{ROMS level 2 `stars` object returned by `get_roms_b10k`}

\item{hindcast}{Weekly means for ROMS level 2 hindcast, returned by `weight_weekly`}

\item{historical}{Weekly means for ROMS level 2 historical run, returned by `weight_weekly`}

\item{lower}{If applicable, lower threshold for returned variable}

\item{upper}{If applicable, upper thresdhold for returned variable}
}
\value{
A `stars` object
}
\description{
Delta-correct ROMS level 2 outputs
}
