% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_ak_coast}
\alias{get_ak_coast}
\title{Alakska coastline shapefile}
\usage{
get_ak_coast(res = c("medium", "high"))
}
\arguments{
\item{res}{resolution ("medium" or "high")}
}
\value{
An "sf" object
}
\description{
Alakska coastline shapefile
}
