% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{get_catch}
\alias{get_catch}
\title{Get catch for a species and (optionally) add in implicit zeroes using haul data.}
\usage{
get_catch(
  species_code,
  zero_expand = FALSE,
  haul_data,
  survey = c("All", "EBS", "SEBS", "NBS", "EBS Slope", "AI", "GOA"),
  years
)
}
\arguments{
\item{species_code}{AKFIN species code returned by `get_species_code` (length 1)}

\item{zero_expand}{Whether to add in implicit zeroes (TRUE) or return only hauls with positive catch (FALSE)}

\item{haul_data}{Haul data returned by `get_hauldata`. If `zero_expand` is TRUE and `haul_data` is missing,
data corresponding to the `survey` argument will be downloaded, however if calling `get_catch` multiple times
for different species, it is more efficient to download the haul data first and pass it to each `get_catch` call.}

\item{survey}{Survey to return data for. Codes differ from the official GAP documentation:
\itemize{
  \item "All" = all surveys
  \item "EBS" = Full Bering Sea Shelf (Southeastern + Northern Bering Sea) survey
  \item "SEBS" = Southeastern Bering Sea survey
  \item "NBS" = Northern Bering Sea survey
  \item "EBS Slope" = Southeastern Bering Sea Slope survey
  \item "GOA" = Gulf of Alaska survey
  \item "AI" = Aleutian Islands survey
}}

\item{years}{Years to subset data to. If missing, returns all years.}
}
\value{
A data.frame
}
\description{
Get catch for a species and (optionally) add in implicit zeroes using haul data.
}
