% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{get_hauldata}
\alias{get_hauldata}
\title{Get data on AFSC survey hauls.}
\source{
<https://afsc-gap-products.github.io/gap_products/content/foss-api-r.html>
}
\usage{
get_hauldata(
  survey = c("All", "EBS", "SEBS", "NBS", "EBS Slope", "AI", "GOA"),
  years
)
}
\arguments{
\item{survey}{Survey to return data for. Codes differ from the official GAP documentation:
\itemize{
  \item "All" = all surveys
  \item "EBS" = Full Bering Sea Shelf (Southeastern + Northern Bering Sea) survey
  \item "SEBS" = Southeastern Bering Sea survey
  \item "NBS" = Northern Bering Sea survey
  \item "EBS Slope" = Southeastern Bering Sea Slope survey
  \item "GOA" = Gulf of Alaska survey
  \item "AI" = Aleutian Islands survey
}}

\item{years}{Years to filter data to. If missing, returns all years.}
}
\value{
A data.frame
}
\description{
Get data on AFSC survey hauls.
}
