% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roms_b10k.R
\name{get_roms_b10k}
\alias{get_roms_b10k}
\title{Download Bering 10K level 2 gridded ROMS output}
\usage{
get_roms_b10k(
  var,
  type = c("hindcast", "projection", "historical"),
  scenario = c("SSP126", "SSP585", "RCP45", "RCP85"),
  start = NA,
  end = NA,
  version = c("K20P19", "K20", "H16"),
  earth_model = c("GFDL", "CESM", "MIROC"),
  crop_ebs = TRUE,
  write_dir = NA
)
}
\arguments{
\item{var}{variable name}

\item{type}{"hindcast", "projection", or "historical" run}

\item{scenario}{"SSP126" or "SSP585" for CMIP6 (`version = "K20P19"` or `version = "K20"`) outputs, "RCP45" or "RCP85" for CMIP5 (`version = "H16"`) outputs}

\item{start}{integer start year. NA defaults to all available years.}

\item{end}{integer end year. NA defaults to all available years.}

\item{version}{Either (1) "K20P19" for 30-layer ROMs model from \insertCite{K20;textual}{BeringSeaData}, with carbonate dynamics from
\insertCite{P19;textual}{BeringSeaData} (default), (2) "K20" for 30-layer model without carbonate dynamics, or (3) "H16" for 10-layer
ROMS model from \insertCite{H16;textual}{BeringSeaData}}

\item{earth_model}{"GFDL", "CESM", or "MIROC"}

\item{crop_ebs}{if TRUE, crop to EBS/NBS survey boundary}

\item{write_dir}{path to write downloaded raster (as GEOTIF). if NA, object is not written to disk.}
}
\value{
A `stars` object
}
\description{
Download Bering 10K level 2 gridded ROMS output
}
\references{
\insertAllCited
}
