% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.R
\name{get_species_codes}
\alias{get_species_codes}
\title{Get AKFIN species data, or return species code corresponding to a species scientific name(s).}
\source{
<https://afsc-gap-products.github.io/gap_products/content/foss-api-r.html>
}
\usage{
get_species_codes(sciname)
}
\arguments{
\item{sciname}{Vector of species scientific names for which to return codes. If missing, returns
all species data. Can be a vector.}
}
\value{
A named vector if sciname is non-missing, otherwise a data.frame.
}
\description{
Get AKFIN species data, or return species code corresponding to a species scientific name(s).
}
