% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_level2_datasets.R
\name{list_level2_datasets}
\alias{list_level2_datasets}
\title{List simulations / datasets available on the thredds server}
\usage{
list_level2_datasets(option = c("sims", "all"), sims = NA, quiet = TRUE)
}
\arguments{
\item{option}{Whether to list simulations ("sims") or all available years / variables for a given
simulation ("all"). If using `option = "all"` with `sims = NA`, it may take a long time to query the
thredds server to list all combinations of available simulations, years, and variables.}

\item{sims}{Optional; if using `option = "all"`, will return a list of all years and variables
available for the provided sim(s)}

\item{quiet}{Set to FALSE to display progress if listing all available years / variables for a given simulation.}
}
\value{
Either a vector of simulation names for `option = "sims"` or a data frame for `option = "all"`
}
\description{
List simulations / datasets available on the thredds server
}
