% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_utilities.R
\name{st_as_stars.tidync}
\alias{st_as_stars.tidync}
\alias{st_as_stars}
\title{Convert tidync object to `stars`}
\source{
<https://github.com/ropensci/tidync/issues/68>
}
\usage{
\method{st_as_stars}{tidync}(.x)
}
\arguments{
\item{.x}{tidync object}
}
\value{
A `stars` object
}
\description{
Convert tidync object to `stars`
}
