% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_utilities.R
\name{st_reband}
\alias{st_reband}
\title{Change multi-band, single attribute `stars` object to single-band, multi-attribute}
\usage{
st_reband(x)
}
\arguments{
\item{x}{stars object with three dimensions}
}
\value{
A stars object
}
\description{
Change multi-band, single attribute `stars` object to single-band, multi-attribute
}
