% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars_utilities.R
\name{st_replicate}
\alias{st_replicate}
\title{Expand a 2-dimensional `stars` object over time}
\usage{
st_replicate(x, name, values = 1, ...)
}
\arguments{
\item{x}{2-dimensional `stars` object}

\item{name}{Name of time dimension}

\item{values}{Values to use for time dimension}
}
\value{
A 3-dimensional `stars` object
}
\description{
Expand a 2-dimensional `stars` object over time
}
