% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ncdf.R
\name{write_ncdf}
\alias{write_ncdf}
\title{Save ROMS or MOM6 output as netcdf}
\usage{
write_ncdf(roms, path, type = c("resampled", "native"), cellsize = 5000)
}
\arguments{
\item{roms}{Object to save (non-cropped)}

\item{path}{File path}

\item{type}{"resampled" for object resampled to a regular UTM zone 2 grid, or "native" for original ROMS grid (currently only implemented for ROMS)}

\item{cellsize}{cell size (in km) used for resampling}
}
\description{
Save ROMS or MOM6 output as netcdf
}
