% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_roms_ncdf.R
\name{write_roms_ncdf}
\alias{write_roms_ncdf}
\title{Save ROMS output as netcdf}
\usage{
write_roms_ncdf(roms, path, type = c("resampled", "native"), cellsize = 5000)
}
\arguments{
\item{roms}{Object to save (non-cropped)}

\item{path}{File path}

\item{type}{"resampled" for object resampled to a regular UTM zone 2 grid, or "native" for original ROMS grid}

\item{cellsize}{cell size (in km) used for resmapling}
}
\description{
Save ROMS output as netcdf
}
